Create a robot and test class for Customize Settings menu
